Pull tagged objects into its own module
